@@ -1,7 +1,5 @@
-#include "mjolnir/ferry_connections.h"
-#include "mjolnir/node_expander.h"
-
 #include <list>
+#include <optional>
 #include <queue>
 #include <unordered_set>
 #include <vector>
@@ -9,7 +7,8 @@
 #include "baldr/graphid.h"
 #include "baldr/json.h"
 #include "midgard/util.h"
-
+#include "mjolnir/ferry_connections.h"
+#include "mjolnir/node_expander.h"
 #include "mjolnir/util.h"
 
 using namespace valhalla::baldr;
@@ -48,7 +47,7 @@ struct LinkGraphNode {
 
 inline bool IsDriveableNonLink(const Edge& edge) {
   return !edge.attributes.link &&
-         (edge.attributes.driveableforward || edge.attributes.driveablereverse) &&
+         ((edge.fwd_access & kAutoAccess) || (edge.rev_access & kAutoAccess)) &&
          edge.attributes.importance != kServiceClass;
 }
 
@@ -116,7 +115,7 @@ nodelist_t FormExitNodes(sequence<Node>& nodes, sequence<Edge>& edges) {
     if (bundle.node.link_edge_ && bundle.node.non_link_edge_) {
       // Check if this node has a link edge that is driveable from the node
       for (const auto& edge : bundle.node_edges) {
-        if (edge.first.attributes.link && edge.first.attributes.driveforward) {
+        if (edge.first.attributes.link && (edge.first.attributes.driveforward)) {
           // Get the highest classification of non-link edges at this node.
           // Add to the exit node list if a valid classification...if no
           // connecting edge is driveable the node will be skipped.
@@ -415,7 +414,9 @@ struct LinkGraphBuilder {
 
     // Make sure that number of children does not exceed the threshold
     if (graph_[from].children.size() >= kMaxLinkEdges) {
-      throw std::runtime_error("Exceeding kMaxLinkEdges in ReclassifyLinks");
+      auto ll = graph_[from].bundle.node.latlng();
+      throw std::runtime_error("Exceeding kMaxLinkEdges in ReclassifyLinks at location " +
+                               std::to_string(ll.lng()) + "," + std::to_string(ll.lat()));
     }
 
     graph_[to].parents.push_back(from);
@@ -521,11 +522,11 @@ private:
 bool IsEdgeDriveableInDirection(uint32_t from_node, const Edge& edge, bool forward) {
   bool right_direction = true;
   if (forward) {
-    right_direction = (edge.sourcenode_ == from_node && edge.attributes.driveableforward) ||
-                      (edge.targetnode_ == from_node && edge.attributes.driveablereverse);
+    right_direction = (edge.sourcenode_ == from_node && (edge.fwd_access & kAutoAccess)) ||
+                      (edge.targetnode_ == from_node && (edge.rev_access & kAutoAccess));
   } else {
-    right_direction = (edge.sourcenode_ == from_node && edge.attributes.driveablereverse) ||
-                      (edge.targetnode_ == from_node && edge.attributes.driveableforward);
+    right_direction = (edge.sourcenode_ == from_node && (edge.rev_access & kAutoAccess)) ||
+                      (edge.targetnode_ == from_node && (edge.fwd_access & kAutoAccess));
   }
   return right_direction;
 }
@@ -641,7 +642,7 @@ bool IsSlipLane(Data& data, SlipLaneInput input, double traverse_threshold) {
       GoTowardsIntersection(input.last_node, input.merge_edge, false, traverse_threshold, data);
 
   // check if two directions intersect
-  boost::optional<uint32_t> intersection_node;
+  std::optional<uint32_t> intersection_node;
   for (auto node : reverse_nodes) {
     if (std::find(forward_nodes.begin(), forward_nodes.end(), node) != forward_nodes.end()) {
       intersection_node = node;
@@ -656,7 +657,7 @@ bool IsSlipLane(Data& data, SlipLaneInput input, double traverse_threshold) {
                                            *intersection_node));
   }
 
-  return intersection_node.is_initialized();
+  return intersection_node != std::nullopt;
 }
 
 SlipLaneInput GetSlipLaneInput(Data& data, const std::vector<uint32_t>& link_edges) {
@@ -692,13 +693,13 @@ SlipLaneInput GetSlipLaneInput(Data& data, const std::vector<uint32_t>& link_edg
   // link_edges store link sequence in reverse order
   // so first link edge is actually the last in the list
   Edge first_link_edge = *data.edges[link_edges.back()];
-  res.first_node = first_link_edge.attributes.driveableforward ? first_link_edge.sourcenode_
-                                                               : first_link_edge.targetnode_;
+  res.first_node = (first_link_edge.fwd_access & kAutoAccess) ? first_link_edge.sourcenode_
+                                                              : first_link_edge.targetnode_;
   res.fork_edge = find_closest_neighbour_edge(res.first_node, first_link_edge, true);
 
   Edge last_link_edge = *data.edges[link_edges.front()];
-  res.last_node = last_link_edge.attributes.driveableforward ? last_link_edge.targetnode_
-                                                             : last_link_edge.sourcenode_;
+  res.last_node = (last_link_edge.fwd_access & kAutoAccess) ? last_link_edge.targetnode_
+                                                            : last_link_edge.sourcenode_;
   res.merge_edge = find_closest_neighbour_edge(res.last_node, last_link_edge, false);
   return res;
 }
